refactor(VolunteerApplyStatusChangeEvent): 패키지 이동
